Switch to more modular approach and add documentation
`game` and `sls` are now require-able modules to allow outside code to use and
configure them. As a result, a new `bin` directory was created, containing only
a script that runs both SLS and game proxy servers in order to mimic the old
behavior.

Currently `node_modules` resides in the root directory so that the dispatcher
can find the modules when loading them. In the future we may want to load them
from the caller's context instead.

As part of this change, `sls` was rewritten to use ES6 features and thus now
requires Node version 4.0.0 or higher.