Convert tera-proxy-game to JS (ES5)
ES6 would make a lot of this code cleaner but seeing as tera-proxy-game
can be run on servers and thus isn't subject to the same requirement as
tera-proxy-sls of Node >=4.0.0 for ES6 features, let's hold off on ES6
for now.

There were also a few minor modifications, including much better stderr
messages for `protocol` and a switch to native `Buffer` methods in the
`encryption` code, the latter of which was not tested for performance
and may need to be reverted pending benchmarks, if necessary.